// Suinocultura-dan/suinocultura/src/app/model/suino.ts
export interface Suino {
  brinco: string;
  brinco_pai: string;
  brinco_mae: string;
  dt_nasc: string;
  dt_saida: string;
  status: string;
  sexo: string;
  pesos: { peso: string; dt_pesagem: string; }[];
}

export interface SuinoPesos {
  peso: number;
  dt_pesagem: string;
  brinco_suino: string;
}